requirements.txt
Django==4.1.6
django-rest-framework==0.1.0
jwt-django